refactor: move ToS / legal links into flavor source sets
The Play-Store-only Terms-of-Use checkbox and the "About & Legal"
Settings section both contain hardcoded GitHub URLs to the published
PRIVACY.md. Previously they were in src/main and gated at call sites
by `BuildConfig.FLAVOR == "play"` — which works at runtime but still
compiles the GitHub URLs into the F-Droid APK.

This commit moves the URL-bearing UI into src/play and adds empty
src/fdroid stubs with the same signatures, so the F-Droid build is
physically free of the URLs.

New composables (same package + signature in both flavor source sets,
so callers in src/main link to whichever flavor is being built):

- com.vitorpamplona.amethyst.ui.screen.loggedOff.legal.TermsGate
  * src/play: renders an AcceptTerms checkbox with the PRIVACY.md link
    and the "acceptance required" error text.
  * src/fdroid: empty body.

- com.vitorpamplona.amethyst.ui.screen.loggedIn.settings.LegalSettingsSection
  * src/play: renders the SettingsSection with Privacy Policy and
    Child Safety Standards rows that open the GitHub-hosted PRIVACY.md.
  * src/fdroid: empty body.

Call-site changes:

- LoginScreen: drop `BuildConfig.FLAVOR == "play"` guard and call
  `TermsGate(...)` inside the existing `isFirstLogin` block. The
  flavor source set picks the right body.
- SignUpScreen: drop `BuildConfig.FLAVOR == "play"` guard, call
  `TermsGate(...)` unconditionally.
- AllSettingsScreen: drop `BuildConfig.FLAVOR == "play"` guard and
  the inline About & Legal SettingsSection, call
  `LegalSettingsSection()` instead. LocalUriHandler and BuildConfig
  imports removed.

The old `src/main/.../loggedOff/AcceptTerms.kt` is deleted; its body
moves into the play-flavor TermsGate as a file-private helper.

The LoginViewModel / SignUpViewModel still use `BuildConfig.FLAVOR` to
decide the initial `acceptedTerms` value (so the login/signup button
isn't disabled on F-Droid). That check is logic only, contains no
URLs, and is safe for F-Droid distribution.

Verified: `grep -r "github.com/vitorpamplona/amethyst/blob"` against
src/main + src/fdroid returns zero matches. Both
:amethyst:compileFdroidDebugKotlin and :amethyst:compilePlayDebugKotlin
compile cleanly.
